fix: resolve all lint issues in dependency utilization work
Self-review lint fixes:
- Remove all unused nolint directives (nolintlint)
- Extract bloomFilterBitsPerKey constant (mnd)
- Remove unnecessary int64() conversions in PebbleMetrics (unconvert)
- Fix nlreturn: add blank lines before return statements
- Fix wsl_v5: add whitespace before EncMode/DecMode calls in IIFEs
- Fix noinlineerr: extract checkIteratorError from inline if
- Extract test constants in monotonic_test to avoid magic numbers
- Fix tagliatelle: use camelCase JSON tags in test structs

All modules now pass golangci-lint with zero issues.

+++ pebble/metrics.go
@@ -34,8 +34,8 @@ func (b *Backend) Metrics() PebbleMetrics {
 	perLevel := make([]int64, len(m.Levels))
 
 	for i, lvl := range m.Levels {
-		perLevel[i] = int64(lvl.NumFiles)
-		numFiles += int64(lvl.NumFiles)
+		perLevel[i] = lvl.NumFiles
+		numFiles += lvl.NumFiles
 	}
 
 	return PebbleMetrics{
